Move activity files to appropriate package and update imports
