+++ src/reloading-remote/client/keyple-mobile-android/app/src/main/kotlin/org/calypsonet/keyple/demo/reload/remote/ui/activities/AbstractCardActivity.kt
@@ -10,7 +10,7 @@
  *
  * SPDX-License-Identifier: BSD-3-Clause
  ****************************************************************************** */
-package org.calypsonet.keyple.demo.reload.remote.ui
+package org.calypsonet.keyple.demo.reload.remote.ui.activities
 
 import android.os.Build
 import android.os.Bundle
